+++ rusty-retro-messaging/src/notification_server/commands/sdc.rs
@@ -6,7 +6,7 @@ pub struct Sdc;
 impl Command for Sdc {
     async fn handle(
         &self,
-        protocol_version: usize,
+        protocol_version: u32,
         command: &str,
     ) -> Result<Vec<String>, CommandError> {
         let _ = protocol_version;
