+++ rusty-retro-messaging/src/notification_server/commands/traits/command.rs
@@ -3,7 +3,7 @@ use crate::errors::command_error::CommandError;
 pub trait Command {
     async fn handle(
         &self,
-        protocol_version: usize,
+        protocol_version: u32,
         command: &str,
     ) -> Result<Vec<String>, CommandError>;
 }
