+++ rusty-retro-messaging/src/notification_server/commands/usr_i.rs
@@ -15,7 +15,7 @@ impl UsrI {
 impl Command for UsrI {
     async fn handle(
         &self,
-        protocol_version: usize,
+        protocol_version: u32,
         command: &str,
     ) -> Result<Vec<String>, CommandError> {
         let _ = protocol_version;
